ユーザーの認証処理を共通化

+++ src/main/java/com/example/iruka_backend/repository/DeckRepository.java
@@ -42,5 +42,5 @@ public interface DeckRepository {
    *
    * @param deckId デッキID
    */
-  int getUserIdByDeckId(Long deckId);
+  Long getUserIdByDeckId(Long deckId);
 }
